Arreglado bug de posición. Archivos de tipografía movidos, leve limpieza

+++ src/utils/Calculador.cpp
@@ -309,7 +309,7 @@ Camino Calculador::obtenerCaminoMin(Escenario *esc, Coordenada coord_pixel_orig,
 
 Coordenada Calculador::generarPosRandom(int size_x_final, int size_x_inicial, int size_y_final, int size_y_inicial, int seed = 0){
 	int x_rand, y_rand;	// Esto FRUSTRA el randomness, siempre va a dar lo mismo, porque vuelve al número múltiplo de 50...
-	srand(((int)time(0)) * SDL_GetTicks()); //seedeo el random bien
+	srand((time(0)) * SDL_GetTicks()); //seedeo el random bien
 	x_rand = (rand() % (size_x_final - size_x_inicial)) + size_x_inicial;
 	y_rand = (rand() % (size_y_final - size_y_inicial)) + size_y_inicial;
